+++ src/main/kotlin/theater/dao/FeatureDao.kt
@@ -6,7 +6,7 @@ import theater.model.*
 
 class FeatureDao(private val dataSource: DataSource) {
 
-    fun createFeature(toCreate: Feature): Long {
+    fun createFeature(toCreate: Feature): Int {
         val stmt = dataSource.connection.prepareStatement(
             "INSERT INTO features (name, value) VALUES (?, ?)",
             Statement.RETURN_GENERATED_KEYS
@@ -17,10 +17,10 @@ class FeatureDao(private val dataSource: DataSource) {
         val gk = stmt.generatedKeys
         gk.next()
 
-        return gk.getLong(1)
+        return gk.getInt(1)
     }
 
-    fun createFeatures(toCreate: Iterable<Feature>): List<Long> {
+    fun createFeatures(toCreate: Iterable<Feature>): List<Int> {
         val stmt = dataSource.connection.prepareStatement(
             "INSERT INTO features (name, value) VALUES (?, ?)",
             Statement.RETURN_GENERATED_KEYS
@@ -34,49 +34,49 @@ class FeatureDao(private val dataSource: DataSource) {
 
         stmt.executeBatch()
         val gk = stmt.generatedKeys
-        val res = ArrayList<Long>()
+        val res = ArrayList<Int>()
         while (gk.next()) {
-            res += gk.getLong(1)
+            res += gk.getInt(1)
         }
 
         return res
     }
 
-    fun findFeature(id: Long): Feature? {
+    fun findFeature(id: Int): Feature? {
         val stmt = dataSource.connection.prepareStatement(
             "SELECT " +
                     "f.id fid, f.name, f.value "+
                     "FROM features AS f " +
                     "WHERE f.id = ?"
         )
-        stmt.setLong(1, id)
+        stmt.setInt(1, id)
         val rs = stmt.executeQuery()
         return if (rs.next()) {
             Feature(
-                rs.getLong("fid"), rs.getString("name"), rs.getString("value"))
+                rs.getInt("fid"), rs.getString("name"), rs.getString("value"))
         } else {
             null
         }
     }
 
-    fun deleteFeature(id: Long): Long {
+    fun deleteFeature(id: Int): Int {
         val stmt = dataSource.connection.prepareStatement(
             "DELETE FROM features WHERE id = ?",
             Statement.RETURN_GENERATED_KEYS
         )
-        stmt.setLong(1, id)
+        stmt.setInt(1, id)
         stmt.executeUpdate()
         val gk = stmt.generatedKeys
         gk.next()
 
-        return gk.getLong(1)
+        return gk.getInt(1)
     }
 
     fun updateFeature(feature: Feature) {
         val stmt = dataSource.connection.prepareStatement("UPDATE features SET name = ?, value = ? WHERE id = ?")
         stmt.setString(1, feature.name)
         stmt.setString(2, feature.value)
-        stmt.setLong(3, feature.id!!)
+        stmt.setInt(3, feature.id!!)
         stmt.executeUpdate()
     }
 
@@ -92,7 +92,7 @@ class FeatureDao(private val dataSource: DataSource) {
         while (rs.next()) {
             res.add(
                 Feature(
-                rs.getLong("id"), rs.getString("name"), rs.getString("value"))
+                rs.getInt("id"), rs.getString("name"), rs.getString("value"))
             )
         }
         return res
